Consolidate semantic tokens test into one file, that uses theories instead of inheritance
