--- src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/SemanticTokenTestBase.cs
@@ -1,223 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Immutable;
-using System.IO;
-using System.Reflection;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Razor.Language;
-using Microsoft.AspNetCore.Razor.LanguageServer.Completion;
-using Microsoft.AspNetCore.Razor.LanguageServer.Extensions;
-using Microsoft.AspNetCore.Razor.LanguageServer.Test.Common;
-using Microsoft.AspNetCore.Razor.PooledObjects;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Razor.Workspaces.Extensions;
-using Microsoft.CodeAnalysis.Text;
-using Microsoft.VisualStudio.LanguageServer.Protocol;
-using Roslyn.Test.Utilities;
-using Xunit;
-using Xunit.Abstractions;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic;
-
-public abstract class SemanticTokenTestBase : TagHelperServiceTestBase
-{
-    private static readonly AsyncLocal<string?> s_fileName = new();
-    private static readonly string s_projectPath = TestProject.GetProjectDirectory(typeof(TagHelperServiceTestBase), layer: TestProject.Layer.Tooling);
-
-    protected static readonly VSInternalServerCapabilities SemanticTokensServerCapabilities = new()
-    {
-        SemanticTokensOptions = new()
-        {
-            Full = false,
-            Range = true
-        }
-    };
-    // Used by the test framework to set the 'base' name for test files.
-    public static string? FileName
-    {
-        get { return s_fileName.Value; }
-        set { s_fileName.Value = value; }
-    }
-
-#if GENERATE_BASELINES
-    protected bool GenerateBaselines { get; set; } = true;
-#else
-    protected bool GenerateBaselines { get; set; } = false;
-#endif
-
-    protected int BaselineTestCount { get; set; }
-    protected int BaselineEditTestCount { get; set; }
-    protected bool UsePreciseSemanticTokenRanges { get; set; }
-
-    protected SemanticTokenTestBase(ITestOutputHelper testOutput, bool usePreciseSemanticTokenRanges)
-        : base(testOutput)
-    {
-        UsePreciseSemanticTokenRanges = usePreciseSemanticTokenRanges;
-    }
-
-    protected void AssertSemanticTokensMatchesBaseline(SourceText sourceText, int[]? actualSemanticTokens)
-    {
-        if (FileName is null)
-        {
-            var message = $"{nameof(AssertSemanticTokensMatchesBaseline)} should only be called from a Semantic test ({nameof(FileName)} is null).";
-            throw new InvalidOperationException(message);
-        }
-
-        var fileName = BaselineTestCount > 0 ? FileName + $"_{BaselineTestCount}" : FileName;
-        var baselineFileName = Path.ChangeExtension(fileName, ".semantic.txt");
-
-        var actualFileContents = GetFileRepresentationOfTokens(sourceText, actualSemanticTokens);
-
-        BaselineTestCount++;
-        if (GenerateBaselines)
-        {
-            GenerateSemanticBaseline(actualFileContents, baselineFileName);
-        }
-
-        var expectedFileContents = GetBaselineFileContents(baselineFileName);
-
-        AssertEx.EqualOrDiff(expectedFileContents, actualFileContents);
-    }
-
-    protected string GetBaselineFileContents(string baselineFileName)
-    {
-        var semanticFile = TestFile.Create(baselineFileName, GetType().GetTypeInfo().Assembly);
-        if (!semanticFile.Exists())
-        {
-            return string.Empty;
-        }
-
-        return semanticFile.ReadAllText();
-    }
-
-    private protected async Task<ProvideSemanticTokensResponse> GetCSharpSemanticTokensResponseAsync(
-        string documentText, Range razorRange, bool isRazorFile)
-    {
-        var codeDocument = CreateCodeDocument(documentText, isRazorFile, DefaultTagHelpers);
-        var csharpDocumentUri = new Uri("C:\\TestSolution\\TestProject\\TestDocument.cs");
-        var csharpSourceText = codeDocument.GetCSharpSourceText();
-
-        await using var csharpServer = await CSharpTestLspServerHelpers.CreateCSharpLspServerAsync(
-            csharpSourceText, csharpDocumentUri, SemanticTokensServerCapabilities, SpanMappingService, DisposalToken);
-
-        var csharpRanges = GetMappedCSharpRanges(codeDocument, razorRange);
-        if (csharpRanges == null)
-        {
-            return new ProvideSemanticTokensResponse(tokens: Array.Empty<int>(), hostDocumentSyncVersion: 0);
-        }
-
-        if (UsePreciseSemanticTokenRanges)
-        {
-            var result = await csharpServer.ExecuteRequestAsync<SemanticTokensRangesParams, SemanticTokens>(
-                "roslyn/semanticTokenRanges",
-                CreateVSSemanticTokensRangesParams(csharpRanges, csharpDocumentUri),
-                DisposalToken);
-
-            return new ProvideSemanticTokensResponse(tokens: result?.Data, hostDocumentSyncVersion: 0);
-        }
-        else
-        {
-            var range = Assert.Single(csharpRanges);
-            var result = await csharpServer.ExecuteRequestAsync<SemanticTokensRangeParams, SemanticTokens>(
-                "textDocument/semanticTokens/range",
-                CreateVSSemanticTokensRangeParams(range, csharpDocumentUri),
-                DisposalToken);
-
-            return new ProvideSemanticTokensResponse(tokens: result?.Data, hostDocumentSyncVersion: 0);
-        }
-    }
-
-    protected Range[]? GetMappedCSharpRanges(RazorCodeDocument codeDocument, Range razorRange)
-    {
-        var documentMappingService = new RazorDocumentMappingService(
-            FilePathService, new TestDocumentContextFactory(), LoggerFactory);
-        if (UsePreciseSemanticTokenRanges)
-        {
-            if (!RazorSemanticTokensInfoService.TryGetSortedCSharpRanges(codeDocument, razorRange, out var csharpRanges))
-            {
-                // No C# in the range.
-                return null;
-            }
-
-            return csharpRanges;
-        }
-
-        Range? range;
-        if (!documentMappingService.TryMapToGeneratedDocumentRange(codeDocument.GetCSharpDocument(), razorRange, out range) &&
-            !RazorSemanticTokensInfoService.TryGetMinimalCSharpRange(codeDocument, razorRange, out range))
-        {
-            // No C# in the range.
-            return null;
-        }
-
-        return [range];
-    }
-
-    internal static SemanticTokensRangesParams CreateVSSemanticTokensRangesParams(Range[] ranges, Uri uri)
-        => new()
-        {
-            TextDocument = new TextDocumentIdentifier { Uri = uri },
-            Ranges = ranges
-        };
-
-    internal static SemanticTokensRangeParams CreateVSSemanticTokensRangeParams(Range range, Uri uri)
-        => new()
-        {
-            TextDocument = new TextDocumentIdentifier { Uri = uri },
-            Range = range
-        };
-
-    private static void GenerateSemanticBaseline(string actualFileContents, string baselineFileName)
-    {
-        var semanticBaselinePath = Path.Combine(s_projectPath, baselineFileName);
-        File.WriteAllText(semanticBaselinePath, actualFileContents);
-    }
-
-    private static string GetFileRepresentationOfTokens(SourceText sourceText, int[]? data)
-    {
-        if (data == null)
-        {
-            return string.Empty;
-        }
-
-        using var _ = StringBuilderPool.GetPooledObject(out var builder);
-        builder.AppendLine("//line,characterPos,length,tokenType,modifier,text");
-        var legendArray = TestRazorSemanticTokensLegend.Instance.Legend.TokenTypes;
-        var prevLength = 0;
-        var lineIndex = 0;
-        var lineOffset = 0;
-        for (var i = 0; i < data.Length; i += 5)
-        {
-            var lineDelta = data[i];
-            var charDelta = data[i + 1];
-            var length = data[i + 2];
-
-            Assert.False(i != 0 && lineDelta == 0 && charDelta == 0, "line delta and character delta are both 0, which is invalid as we shouldn't be producing overlapping tokens");
-            Assert.False(i != 0 && lineDelta == 0 && charDelta < prevLength, "Previous length is longer than char offset from previous start, meaning tokens will overlap");
-
-            if (lineDelta != 0)
-            {
-                lineOffset = 0;
-            }
-
-            lineIndex += lineDelta;
-            lineOffset += charDelta;
-
-            var typeString = legendArray[data[i + 3]];
-            builder.Append(lineDelta).Append(' ');
-            builder.Append(charDelta).Append(' ');
-            builder.Append(length).Append(' ');
-            builder.Append(typeString).Append(' ');
-            builder.Append(data[i + 4]).Append(' ');
-            builder.Append('[').Append(sourceText.GetSubTextString(new TextSpan(sourceText.Lines[lineIndex].Start + lineOffset, length))).Append(']');
-            builder.AppendLine();
-
-            prevLength = length;
-        }
-
-        return builder.ToString();
-    }
-}
